Hard coded English t function working. Separate HOCs for Client and Server..
