+++ lib/i18nClient.js
@@ -10,13 +10,10 @@ const mapStateToProps = reduxState => {
 const withI18N = WrappedComponent => {
   class WithI18N extends Component {
     t = key => {
-      if (this.props.__NEXT_DATA__) {
-        var translations = this.props.__NEXT_DATA__.props.initialState
-          .translations;
-        for (var p in translations) {
-          if (key == translations[p].key || key == translations[p].id)
-            return translations[p].English;
-        }
+      let translations = this.props.translations;
+      for (var p in translations) {
+        if (key == translations[p].key || key == translations[p].id)
+          return translations[p].English;
       }
       return key;
     };
@@ -24,7 +21,7 @@ const withI18N = WrappedComponent => {
       return <WrappedComponent {...this.props} t={this.t} />;
     }
   }
-  return WithI18N;
+  return connect(mapStateToProps)(WithI18N);
 };
 
 export default withI18N;
